requirements.txt
streamlit
google-generativeai
python-dotenv
validators
fpdf2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2